+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.2] - 2022-01-30
+
+### Fixed
+- Bug where contact phone and contact address were always assumed to be present
+
+## [1.4.1] - 2022-01-30
+
+### Added
+- More verbose logging
+
 ## [1.4.0] - 2022-01-12
 
 ### Changed
@@ -64,6 +74,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Check authentication attributes before interacting with API.
 
 
+[1.4.1]: https://github.com/mujde-aze/nt-postman-dt-proxy/compare/v1.4.1...v1.4.2
+[1.4.1]: https://github.com/mujde-aze/nt-postman-dt-proxy/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/mujde-aze/nt-postman-dt-proxy/compare/v1.3.4...v1.4.0
 [1.3.4]: https://github.com/mujde-aze/nt-postman-dt-proxy/compare/v1.3.3...v1.3.4
 [1.3.3]: https://github.com/mujde-aze/nt-postman-dt-proxy/compare/v1.3.2...v1.3.3